move check_tlp_marking to markings\utils.py
